requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
aiohttp==3.9.1  # Hỗ trợ gọi API không đồng bộ

# LLM và xử lý ngôn ngữ
llama-cpp-python==0.2.19
langchain==0.0.335
langchain-community==0.0.13
sentence-transformers==2.2.2
faiss-cpu==1.7.4  

# Xử lý văn bản
regex==2023.10.3
unidecode==1.3.7
nltk==3.8.1

# Hỗ trợ CUDA
nvidia-ml-py==12.535.108  # Để kiểm tra GPU status

# Utilities
numpy==1.26.1
pandas==2.1.2
tqdm==4.66.1

# Debug và logging
rich==13.6.0
loguru==0.7.2